Move the test to its own package
